GetAnimationTotalTime int to float

+++ Engine/Component/ComponentAnimation.cpp
@@ -126,7 +126,7 @@ ENGINE_API float ComponentAnimation::GetCurrentClipPercentatge() const
 	return 0.0f;
 }
 
-ENGINE_API int ComponentAnimation::GetTotalAnimationTime() const
+ENGINE_API float ComponentAnimation::GetTotalAnimationTime() const
 {
 	for (auto& playing_clip : animation_controller->playing_clips)
 	{
@@ -138,7 +138,7 @@ ENGINE_API int ComponentAnimation::GetTotalAnimationTime() const
 		return playing_clip.clip->animation_time;
 
 	}
-	return 0;
+	return 0.0f;
 }
